requirements.txt
# Dashboard dependencies
requests>=2.31.0
pyyaml>=6.0.1
jinja2>=3.1.2
python-dateutil>=2.8.2

# Chart generation dependencies
matplotlib>=3.8.0
numpy>=1.26.0
scipy>=1.11.0
seaborn>=0.13.0
pandas>=2.1.0
